Add files via upload
These are just snapshots of the winding machine that I just took... That black box under the monitor screen is where that Raspberry-Pi 4 resides. There are other electronic components inside that box, in order to establish different operating voltages for different devices. The 2 arrow signs seen on the front side of the box indicates the wire-guide direction. This machine is actually still in a crude format, and it still needs to have a few more buttons to expand controllability over the machine.. So far, there are buttons to Start/Pause the machine during a process, ABORT the process, change wire-guide direction, as well as adjust the wire-guide position manually while on pause.